Initial contribution of dash-licenses-wrapper
This contribution provides a wrapper, to make it easier to integrate and run
dash-licenses, in Eclipse Foundation projects repositories. It's mainly aiming
to cover JavaScript and TypeScript projects, but might be usable in others too.

Once published to `npm`, the wrapper can be installed as a devDependency and
used to run license checks locally on a developer's laptop and also as part
of GitHub CI, using a provided example `workflow` file.

See README.md for more details

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>